Struct gosub_engine::html5_parser::input_stream::InputStream

pub struct InputStream {
    pub encoding: Encoding,
    pub confidence: Confidence,
    pub position: Position,
    pub length: usize,
    pub has_read_eof: bool,
    /* private fields */
}

HTML(5) input stream structure

Fields
encoding: Encoding

Current encoding

confidence: Confidence

How confident are we that this is the correct encoding?

position: Position

Current positions

length: usize

Length (in chars) of the buffer

has_read_eof: bool

If all things are ok, both buffer and u8_buffer should refer to the same memory location (?)

Implementations

impl InputStream

pub fn new() -> Self

Create a new default empty input stream

pub fn is_certain_encoding(&self) -> bool

Returns true when the encoding encountered is defined as certain

pub fn detect_encoding(&self)

Detect the given encoding from stream analysis

pub fn eof(&self) -> bool

Returns true when the stream pointer is at the end of the stream

pub fn reset(&mut self)

Reset the stream reader back to the start

pub fn seek(&mut self, mode: SeekMode, offset: isize)

Seek explicit offset in the stream (based on chars)

pub fn get_previous_position(&mut self) -> Position

pub fn tell(&self) -> usize

pub fn set_confidence(&mut self, c: Confidence)

Set the given confidence of the input stream encoding

pub fn set_encoding(&mut self, e: Encoding)

Changes the encoding and if necessary, decodes the u8 buffer into the correct encoding

pub fn force_set_encoding(&mut self, e: Encoding)

Sets the encoding for this stream, and decodes the u8_buffer into the buffer with the correct encoding.

pub fn read_from_file(&mut self, f: File, e: Option<Encoding>) -> Result<()>

Populates the current buffer with the contents of given file f

pub fn read_from_str(&mut self, s: &str, e: Option<Encoding>)

Populates the current buffer with the contents of the given string s